apps/users-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
kafka-python==2.0.2
pydantic==2.10.3
email-validator==2.2.0
